added PlatformIO Set up, updated Arduino documentation and code accordingly
