+++ project/SYSTEM.CFG
@@ -2,7 +2,7 @@
 # * Baah Box Arduino : System configuration file *
 # ************************************************
 
-# Copyright (C) 2017 – 2019 Orange SA
+# Copyright (C) 2017 – 2023 Orange SA
 
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
