+++ project/platformio.ini
@@ -1,0 +1,22 @@
+; PlatformIO Project Configuration File
+;
+;   Build options: build flags, source filter
+;   Upload options: custom upload port, speed and extra flags
+;   Library options: dependencies, extra library storages
+;   Advanced options: extra scripting
+;
+; Please visit documentation for the other options and examples
+; https://docs.platformio.org/page/projectconf.html
+
+[env:adafruit_feather_m0]
+platform = atmelsam
+board = adafruit_feather_m0
+framework = arduino
+lib_deps = 
+	arduino-libraries/SD@^1.2.4
+	adafruit/Adafruit SSD1306@^2.5.7
+	adafruit/Adafruit BluefruitLE nRF51@^1.10.0
+	adafruit/Adafruit GFX Library@^1.11.5
+
+[platformio]
+description = PlatformIO version for BaahBoxArduino
